add WIP arguments for number input widget in scriptdialog.cpp
