+++ src/tiled/scriptmanager.cpp
@@ -141,7 +141,8 @@ ScriptManager::ScriptManager(QObject *parent)
     qRegisterMetaType<QPushButton*>();
     qRegisterMetaType<QDialog*>();
     qRegisterMetaType<ColorButton*>();
-
+    qRegisterMetaType<NumberInputArgs>();
+    qRegisterMetaType<NumberInputArgs*>();
     connect(&mWatcher, &FileSystemWatcher::pathsChanged,
             this, &ScriptManager::scriptFilesChanged);
